CBR == RAR != Zip
- Immediately discovered that .cbr is based on RAR, not ZIP. Installed
libraries rarfile and unrar to handle it.

-These are dependent on RarLabs' "unrar.exe". Required an installation
to my machine, and adding to Path. After a lot of fudging, it worked
after adding it to an already-added directory in the 'System variable'
Path (not 'User variable), then pip installing unrar and rarfile.

- It also worked when adding unrar.exe to the current working directory.
I imagine this is the more likely way to go about it if I expect it to
be used on other's machines.
-unrar.exe added to this commit for posterity, but needs to be added to
testing directory (a subfolder currently gitIgnored) for actual testing.

https://rarfile.readthedocs.org/en/latest/faq.html#id5