Installation of this new library
https://github.com/miebach/lib-rar-python/pull/2

This is an attempt by skozacik to convert the previous miebach library
to work with Windows. My personal build of the distribution is included
in backup\skozacik windows\ but I didn't get it to install from it.
After cloning his branch, I did the following instead.

pip uninstall librar (0.0.4)
<move to directory of clone of skoz' branch/pull>
python setup.py install