Rename zip to cbz; Parse attempt
Renaming requires a delete of the old (os.remove) to practically
overwrite it with the rename (os.rename). That's due to some Windows
limitation with Python.

Line 67 is an attempt at parsing out unwanted images from the comic. The
problem is that 'filename' is an object of some type, when it needs to
be a string in order to use that type's Parse function. Line 66 is the
stopgap.