Folder before images issue attempt
NOTE: Further testing needs a folder before the images

Trying to figure out how to get to the images if they are in a folder in
the archive. os.walk seems to be the way to go, but I don't know how to
use its tuple to get to the images.