Folder before images solution; os.walk
Using os.walk to go through the folder. If the images are in a
subfolder, changes current working directory to that folder. Surely
there's a better way to do this, but it works for now. Then for each
file (image) in the current directory, if it passes the filter, it is
written to the zip. Commented out the old solution.